- 'bypass' keyword logic moved to VerdictContext  - VerdictStatement uses VerdictContext to execute queries.
