+++ src/main/java/org/verdictdb/VerdictContext.java
@@ -25,6 +25,7 @@ import org.verdictdb.connection.JdbcConnection;
 import org.verdictdb.coordinator.ExecutionContext;
 import org.verdictdb.coordinator.VerdictResultStream;
 import org.verdictdb.coordinator.VerdictSingleResult;
+import org.verdictdb.coordinator.VerdictSingleResultFromDbmsQueryResult;
 import org.verdictdb.core.scrambling.ScrambleMetaSet;
 import org.verdictdb.exception.VerdictDBDbmsException;
 import org.verdictdb.exception.VerdictDBException;
@@ -235,6 +236,19 @@ public class VerdictContext {
       String originalSchema, String originalTable, String newSchema, String newTable) {}
 
   /**
+   * Check whether given sql contains 'bypass' keyword at the beginning
+   *
+   * @param sql original sql
+   * @return without 'bypass' keyword if the original sql begins with it. null otherwise.
+   */
+  private String checkBypass(String sql) {
+    if (sql.trim().toLowerCase().startsWith("bypass")) {
+      return sql.trim().substring(6);
+    }
+    return null;
+  }
+
+  /**
    * Returns a reliable result set as an answer. Right now, simply returns the first batch of
    * Continuous results.
    *
@@ -246,10 +260,15 @@ public class VerdictContext {
    * @throws VerdictDBException
    */
   public VerdictSingleResult sql(String query) throws VerdictDBException {
-    ExecutionContext exec = createNewExecutionContext();
-    VerdictSingleResult result = exec.sql(query);
-    removeExecutionContext(exec);
-    return result;
+    String bypassSql = checkBypass(query);
+    if (bypassSql != null) {
+      return executeAsIs(bypassSql);
+    } else {
+      ExecutionContext exec = createNewExecutionContext();
+      VerdictSingleResult result = exec.sql(query);
+      removeExecutionContext(exec);
+      return result;
+    }
   }
 
   /**
@@ -262,5 +281,9 @@ public class VerdictContext {
     ExecutionContext exec = createNewExecutionContext();
     VerdictResultStream stream = exec.streamsql(query);
     return stream;
+  }
+
+  private VerdictSingleResult executeAsIs(String sql) throws VerdictDBDbmsException {
+    return new VerdictSingleResultFromDbmsQueryResult(conn.execute(sql));
   }
 }
